requirements.txt
requests>=2.32
pandas>=2.2
numpy>=1.26
plotly>=5.22
matplotlib>=3.8
scikit-learn>=1.4
statsmodels>=0.14
pydantic>=2.7
PyYAML>=6.0
streamlit>=1.37
python-dateutil>=2.9
python-dotenv>=1.0
# Optional (auto ARIMA); if install fails on Windows, the app falls back automatically:
pmdarima>=2.0 ; platform_system != "Windows" or python_version>="3.11"
# Optional (Prophet). On Windows this may require extra toolchain; app works without it:
prophet>=1.1 ; platform_system != "Windows"